requirements.txt
google-adk==1.2.1
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
requests>=2.32.0
python-dotenv>=1.0.0
pydantic>=2.0.0
